decouple combobox's core functions from tag related functions ref #59
